Tidy up by moving to new file.
